+++ src/main/java/com/group1/fmobile/controller/guest/GuestFilterController.java
@@ -3,15 +3,16 @@ package com.group1.fmobile.controller.guest;
 import com.group1.fmobile.domain.Product;
 import com.group1.fmobile.service.ProductServices;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.math.BigDecimal;
+import java.util.*;
 
 @Controller
+@RequestMapping("/guest")
 public class GuestFilterController {
     @Autowired
     public GuestFilterController(ProductServices productServices) {
@@ -42,4 +43,6 @@ public class GuestFilterController {
         model.addAttribute("products", listMobileProducts);
         return "guest/searchPage/fragments/productList";
     }
+
 }
+
