+++ src/main/java/com/group1/fmobile/controller/guest/GuestSearchPageController.java
@@ -3,15 +3,18 @@ package com.group1.fmobile.controller.guest;
 import com.group1.fmobile.domain.Product;
 import com.group1.fmobile.service.ProductServices;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 @Controller
+@RequestMapping("/guest")
 public class GuestSearchPageController {
     @Autowired
     public GuestSearchPageController(ProductServices productServices){
@@ -28,15 +31,18 @@ public class GuestSearchPageController {
         }
 
         List<Product> results = productServices.searchByQuery(query.trim());
+
         if (results.isEmpty()) {
             return "guest/searchPage/noDataFound";
         }
 
         Map<String, Integer> categoryCounts = productServices.countProductsByCategory(results);
+
         model.addAttribute("query", query);
         model.addAttribute("categoryCounts", categoryCounts);
         model.addAttribute("products", results);
-
         return productServices.determineViewName(categoryCounts.keySet());
     }
+
 }
+
